+++ FOX/ff/bonded_calculate.py
@@ -10,13 +10,15 @@ A module for calculating bonded interactions using harmonic + cosine potentials.
 
     V_{angles} = k_{\theta} (\theta - \theta_{0})^2
 
-    V_{diehdrals} = k_{\phi} [1 + \cos(n \phi - \delta)]
+    V_{Urey-Bradley} = k_{\hat{r}} (|hat{r} - \hat{r}_{0})^2
+
+    V_{dihedrals} = k_{\phi} [1 + \cos(n \phi - \delta)]
 
     V_{impropers} = k_{\omega} (\omega - \omega_{0})^2
 
 """
 
-from typing import Union
+from typing import Union, Tuple, Optional
 from itertools import permutations
 
 import numpy as np
@@ -24,6 +26,7 @@ import pandas as pd
 
 from scm.plams import Units
 
+from .parse_wildcards import parse_wildcards
 from ..classes.multi_mol import MultiMolecule
 from ..io.read_psf import PSFContainer
 from ..io.read_prm import PRMContainer
@@ -31,9 +34,8 @@ from ..io.read_prm import PRMContainer
 __all__ = ['get_bonded']
 
 
-def get_bonded(mol: Union[str, MultiMolecule],
-               psf: Union[str, PSFContainer],
-               prm: Union[str, PRMContainer]) -> pd.DataFrame:
+def get_bonded(mol: Union[str, MultiMolecule], psf: Union[str, PSFContainer],
+               prm: Union[str, PRMContainer]) -> Tuple[Optional[pd.DataFrame], ...]:
     r"""Collect forcefield parameters and calculate all intra-ligand interactions in **mol**.
 
     Forcefield parameters are collected from the provided **psf** and **prm** files.
@@ -53,9 +55,9 @@ def get_bonded(mol: Union[str, MultiMolecule],
 
     Returns
     -------
-    4x :class:`pandas.Series` and/or ``None``
-        Four series with the potential energies of all bonds, angles, proper and
-        improper dihedral angles.
+    5x :class:`pandas.Series` and/or ``None``
+        Four series with the potential energies of all bonds, angles, Urey-Bradley terms,
+        proper and improper dihedral angles.
         A Series is replaced with ``None`` if no parameters are available for that particular
         section.
         Units are in atomic units.
@@ -77,31 +79,41 @@ def get_bonded(mol: Union[str, MultiMolecule],
     else:
         mol = mol.copy(deep=False)
     mol.atoms = psf.to_atom_dict()
+    symbols = sorted(mol.atoms.keys())
 
     # Extract parameters from the .prm file
-    bonds, angles, dihedrals, impropers = process_prm(prm)
+    bonds, angles, urey_bradley, dihedrals, impropers = process_prm(prm)
 
     # Calculate the various potential energies
     if bonds is not None:
+        parse_wildcards(bonds, symbols, prm_type='bonds')
         set_V_bonds(bonds, mol, psf.bonds)
         bonds = bonds['V'] * Units.conversion_ratio('kcal/mol', 'au')
 
     if angles is not None:
+        parse_wildcards(bonds, symbols, prm_type='angles')
         set_V_angles(angles, mol, psf.angles)
         angles = angles['V'] * Units.conversion_ratio('kcal/mol', 'au')
 
+    if urey_bradley is not None:
+        parse_wildcards(bonds, symbols, prm_type='urey_bradley')
+        set_V_UB(urey_bradley, mol, psf.angles)
+        urey_bradley = urey_bradley['V'] * Units.conversion_ratio('kcal/mol', 'au')
+
     if dihedrals is not None:
+        parse_wildcards(bonds, symbols, prm_type='dihedrals')
         set_V_dihedrals(dihedrals, mol, psf.dihedrals)
         dihedrals = dihedrals['V'] * Units.conversion_ratio('kcal/mol', 'au')
 
     if impropers is not None:
+        parse_wildcards(bonds, symbols, prm_type='impropers')
         set_V_impropers(impropers, mol, psf.impropers)
         impropers = impropers['V'] * Units.conversion_ratio('kcal/mol', 'au')
 
-    return bonds, angles, dihedrals, impropers
+    return bonds, angles, urey_bradley, dihedrals, impropers
 
 
-def process_prm(prm: Union[PRMContainer, str]):
+def process_prm(prm: Union[PRMContainer, str]) -> tuple:
     """Extract all bond, angle, dihedral and improper parameters from **prm**."""
     if not isinstance(prm, PRMContainer):
         prm = PRMContainer.read(prm)
@@ -115,6 +127,15 @@ def process_prm(prm: Union[PRMContainer, str]):
 
     angles = prm.angles
     if angles is not None:
+        urey_bradley = angles[[5, 6]].copy()
+        urey_bradley.index = urey_bradley.index.droplevel(1)
+        is_null = urey_bradley.isnull()
+        if is_null.values.all():
+            urey_bradley = None
+        else:
+            urey_bradley[is_null] = 0.0
+            urey_bradley['V'] = np.nan
+
         angles = angles[[3, 4]].copy()
         angles[4] *= np.radians(1)
         angles['V'] = np.nan
@@ -131,7 +152,7 @@ def process_prm(prm: Union[PRMContainer, str]):
         impropers[6] *= np.radians(1)
         impropers['V'] = np.nan
 
-    return bonds, angles, dihedrals, impropers
+    return bonds, angles, urey_bradley, dihedrals, impropers
 
 
 def set_V_bonds(df: pd.DataFrame, mol: MultiMolecule, bond_idx: np.ndarray) -> None:
@@ -182,6 +203,32 @@ def set_V_angles(df: pd.DataFrame, mol: MultiMolecule, angle_idx: np.ndarray) ->
         j = np.all(symbol[angle_idx] == i, axis=1)
         j |= np.all(symbol[angle_idx[:, ::-1]] == i, axis=1)  # Consider all valid permutations
         df.at[i, 'V'] = get_V_harmonic(angle[:, j], *item).sum()
+
+
+def set_V_UB(df: pd.DataFrame, mol: MultiMolecule, angle_idx: np.ndarray) -> None:
+    """Calculate and set :math:`V_{Urey-Bradley}` in **df**.
+
+    Parameters
+    ----------
+    df : :class:`pd.DataFrame`
+        A DataFrame with atom pairs and parameters.
+
+    mol : :class:`MultiMolecule`
+        A MultiMolecule instance.
+
+    angle_idx : :math:`(i,3)` :class:`numpy.ndarray`
+         A 2D numpy array with all atom-pairs defining angles.
+
+    """
+    symbol = mol.symbol
+    bond_idx = angle_idx[:, 0::2]
+    distance = _dist(mol, bond_idx)
+
+    iterator = df.iloc[:, 0:2].iterrows()
+    for i, item in iterator:
+        j = np.all(symbol[bond_idx] == i, axis=1)
+        j |= np.all(symbol[bond_idx[:, ::-1]] == i, axis=1)  # Consider all valid permutations
+        df.at[i, 'V'] = get_V_harmonic(distance[:, j], *item).sum()
 
 
 def set_V_dihedrals(df: pd.DataFrame, mol: MultiMolecule, dihed_idx: np.ndarray) -> None:
@@ -309,5 +356,5 @@ def get_V_cos(phi: np.ndarray, k: float, n: int, delta: float = 0.0) -> float:
         The phase-correction :math:`\delta`; units should be in radian.
 
     """  # noqa
-    V = k * np.cos(n * phi - delta)
+    V = k * (1 + np.cos(n * phi - delta))
     return V.mean(axis=0)